requirements.txt
Django==2.2
pytz==2020.1
pyotp==2.4.1
djangorestframework==3.11.0
gunicorn==20.0.4
django-cors-headers==3.3.0
djangorestframework-jwt==1.11.0
